Disable DSDA property which enables QC implementation.
All the QC implementation for DSDA is deprecated due to
limitaitons in support of DSDA functionality from telephony side.
So, handling of use cases like redial, select phone account
is not there in QC side.
So, all the QC execution, which is enabled through
the property by default, needs to be disabled so that, it will
take the default implementation.

CRs-Fixed: 4109792
Change-Id: I1b1cd774399a7eab524b3efd3f2f1c901507a7bb