Delay added to send CLCC response while IMS conference call process
In the process of merging IMS calls, we receive AT+CLCC from the remote while the call list is getting updated to the expected call list of parent conference calls and its child calls.

The IMS merge process follows disconnection of already present calls and creation of those calls as children to the parent conference call.
This process takes a significant time of around 1.5-2 seconds.

This results in empty response for AT+CLCC, making the remote go in bad state as expected response is not met.
Adding this delay will let the callList update in background and send correct response to the remote for AT+CLCC.

CRs-Fixed: 4103533
Change-Id: I5be4574644bdefa5bf19a3d9b2118c4c5e62b4eb