ADM: Clear previous active device in Dual mode scenario on new BT active
When a CSIP device is made active, we need to reset the earlier
HFP and A2dp active devices as null.
When done SHO from a BR EDR device to CSIP device,
due to non clearance of active device null for BR edr profiles,
the sessions are still active and mm is not aware of the devices
added and removed, causing the mismatch and leading to no audio.

CRs-Fixed: 4121462
Change-Id: I1b1cd774399a7eef524b3ab59f1f2c901507a7aa