VOIP WAR: reject startSco request if broadcasted inactive HFP device
In some SHO corner case, SCO get connected in BT host stack, but HFP
active device is still null in framework, so that BT_SCO=on can not
be setting and audio will not route to BT SCO device.

To resolve this issue, reject startSco request if broadcasted HFP device
inactive.

CRs-Fixed: 4485548
Change-Id: Iaace8758c123374b0ee83e700c9ea503badfe008