HFP: Fixed the ANR issue caused due to deadlock of mutexes.
ANR happened in cleanup because of waiting of lock mutex,
which is already acquired by subsciption change listener for
stopping and starting the phonestate change listener.

Removed the dependency of same lock mutex in cleanup and
subscription change as it is not required to use in later
function since the phonestate change listeners uses its own
lock.

CRs-Fixed: 4267085
Change-Id: Icb4c64a7c7742e55a1dbe1e3d110426e5f134dcf